Rename MatesGM1_v17.html to index.html
This is a simple game to learn the multiplication tables. You can choose between three levels of difficulty which are the tables to study and choose what object to blow up: a ballon, an asteroid or a tank.
There is no violence in this game. 
The idea is to learn multiplication replicating what children like to play.